remote directories missing is not an error
When trying to delete a non-existing remote directory, WebDAV returns a
207 so-called multistatus message for potentially several operations.
Until now these message were treated as errors in the final else clause
in line 52.  2xx actually indicate success and need specific treatment.

This commit adds some initial support for multistatus messages. The
response body contains an XML document from which we extract the
```D:status``` element which contains the actual HTTP status message
(such as ```404 Not found```). We inspect this status message and if
it’s a 404 we choose to not handle this as an error because remote
missing directories are actually quite normal.